old_version/gm_general_chat/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
websockets==12.0

# LlamaIndex - Stable versions that work together
llama-index-core==0.10.55
llama-index-llms-azure-openai==0.1.6
llama-index-llms-openai==0.1.24
llama-index-embeddings-azure-openai==0.1.7
llama-index-embeddings-openai==0.1.9

# Redis and utilities
redis==5.0.1
python-dotenv==1.0.0

# YAML support for prompts
PyYAML==6.0.1

# HTTP client for testing and medical tools
aiohttp==3.9.1

# Additional dependencies that might be needed
typing-extensions>=4.8.0
tiktoken